rename files and cleanup
