+++ packages/compass-telemetry/src/generic-track.ts
@@ -62,7 +62,6 @@ export const createTrack = ({
       parameters.connection_id = connectionInfo.id;
     }
 
-    console.log('TRACK', event, parameters);
     sendTrack(event, parameters || {});
   };
